change database transact table

+++ template.sql
@@ -13,7 +13,7 @@ result VARCHAR(50),
 price VARCHAR(30),
 size VARCHAR(30),
 amount INT,
-value REAL);
+value double precision);
 
 
 CREATE TABLE strat(
